update all *.py files
