+++ rapster/binary_evolution.py
@@ -18,8 +18,8 @@
 
 from constants import *
 from functions import *
+from remnant import *
 
-#@njit
 def evolve_BBHs(seed, t, z, dt, zCl_form, binaries, hardening, mergers, mBH, sBH, gBH, n_star, v_star, vBH, t_rlx, m_avg, mBH_avg, na_BH, nc_BH, N_BH, N_BBH, N_me, N_me2b, N_3cap, N_meFi, N_meRe, N_meEj, N_dis, N_ex, N_BHej, N_BBHej, N_hardening, Vc_BH, N_bb, triples, N_Triples):
     """
     @in seed: simulation seed number
