+++ rapster/triples.py
@@ -18,8 +18,8 @@
 
 from constants import *
 from functions import *
+from remnant import *
 
-#@njit
 def evolve_triples(seed, t, z, zCl_form, triples, binaries, mBH, sBH, gBH, mBH_avg, N_Triples, N_BBH, N_BH, N_me, N_meRe, N_meEj, N_ZLK, v_star, vBH, nc_BH, mergers):
     """
     @in seed: simulation seed number
