+++ rapster/two_body_capture.py
@@ -18,8 +18,8 @@
 
 from constants import *
 from functions import *
+from remnant import *
 
-#@njit
 def two_body_capture(seed, t, dt, z, zCl_form, k_2cap, mBH_avg, binaries, mBH, sBH, gBH, vBH, v_star, N_2cap, N_BH, N_BBH, N_me, N_meRe, N_meEj, mergers):
     """
     @in seed: simulation seed number
